Ambiotica port: use float libm (wasm toolchain lacks double exp / lrintf)
Plinky's wasm toolchain provides expf/powf/etc. but not double exp() or lrintf,
which failed at link. Convert the only real uses to float:
- bloom.c, drift.c: exp() -> expf() (float coefficient math).
- looper.c int16 store: lrintf() -> manual round-to-nearest.
Harness audio verified unchanged (6.78 MB, peak 0.352, 0 clip).

+++ ambiotica_port/dsp/bloom.c
@@ -31,7 +31,7 @@ struct bloom_s {
 /* One-pole coefficient for a time constant (a = exp(-1/(t*sr))). */
 static float coef_for(float t_seconds, double sr) {
     if (t_seconds <= 0.0f) return 0.0f;
-    return (float) exp(-1.0 / ((double) t_seconds * sr));
+    return expf(-1.0f / ((float) t_seconds * (float) sr));   /* expf: wasm toolchain has no double exp */
 }
 
 bloom_t* bloom_create(double sample_rate) {
